requirements.txt
api==0.0.7
APIResponse==0.2
asgiref==3.4.1
automium==0.2.6
automium-web==0.1.1
beautifulsoup4==4.9.3
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==8.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
decorator==5.0.9
defusedxml==0.7.1
Django==3.2.5
django-crispy-forms==1.12.0
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
drf-json-api==0.1.1
Flask==2.0.1
idna==3.2
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.1
legacy==0.1.6
MarkupSafe==2.0.1
nose==1.3.7
null==0.6.1
oauthlib==3.1.1
psycopg2==2.9.1
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
self==2020.12.3
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.2.1
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
Werkzeug==2.0.1
